Fixed a potential memory leak caused by a setState order of operations. This was showing as a 'no-op' warning by React.
Added extra validation steps to consumeEndpoint in both CodewordOutput and CodewordInput to prevent firing a request with an empty Codeword or invalid Codeword.

Changed the Generate and Validate buttons to re-enforce feedback on invalid input, At the moment by disabling the button, this may change to recolouring the button in the future.

Fixed minor over-site with how steps was implemented in CodewordInput